+++ src/core/mod.rs
@@ -7,6 +7,7 @@ pub mod msg;
 
 pub mod err;
 mod save_struct_macro;
+pub mod song_order;
 
 pub use err::*;
 //pub use save_struct_macro::*;